Handle vcluster config for user add (#102)
* Handle vcluster config for user add

* typing hack

* fixed formatting

* changes

* Working changes

* cleanup

* Update test command permissions and refactor cluster display components

This commit modifies the test command in `package.json` to include `--allow-all` permissions for Deno. Additionally, it refactors the `ClusterDisplay` and `ClusterRowWithContracts` components in `clusters.tsx` for improved readability and consistency. The `K8sCredential` type is enhanced in `credentialTypes.ts`, and various minor formatting adjustments are made across several files for better code clarity.

---------

Co-authored-by: John Pham <johnphammail@gmail.com>